test(banner): verify @ttsc/banner applies to a ttsx-compiled dependency
ttsx emits each dependency through its own plugin host now, so a dependency that
configures @ttsc/banner gets the banner. The old tests asserted the removed
per-package on-disk emit cache; they now capture the dependency's emitted
JavaScript through a `load` hook the entry installs (registered after ttsx's, so
it observes ttsx's output) and assert the banner is present.

- applies-when-ttsx-compiles-a-dependency: the captured dependency emit carries
  the banner.
- reflects-a-dependency-config-change (was rebuilds-...-cache-when-config-changes):
  editing banner.config is reflected on the next run via fresh emit, no cache.
